+++ apiserver/facades/agent/provisioner/containerizer_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -typed -package mocks -destination mocks/containerizer_mock.go github.com/juju/juju/internal/network/containerizer LinkLayerDevice
+//	mockgen -typed -package provisioner -destination containerizer_mock_test.go github.com/juju/juju/internal/network/containerizer LinkLayerDevice
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package provisioner is a generated GoMock package.
+package provisioner
 
 import (
 	reflect "reflect"
