+++ apiserver/facades/agent/provisioner/interface_mock_test.go
@@ -3,17 +3,16 @@
 //
 // Generated by this command:
 //
-//	mockgen -typed -package mocks -destination mocks/package_mock.go github.com/juju/juju/apiserver/facades/agent/provisioner Machine,BridgePolicy,Unit,Application,Charm
+//	mockgen -typed -package provisioner -destination interface_mock_test.go github.com/juju/juju/apiserver/facades/agent/provisioner Machine,BridgePolicy,Unit,Application,Charm
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package provisioner is a generated GoMock package.
+package provisioner
 
 import (
 	reflect "reflect"
 
 	set "github.com/juju/collections/set"
-	provisioner "github.com/juju/juju/apiserver/facades/agent/provisioner"
 	constraints "github.com/juju/juju/core/constraints"
 	instance "github.com/juju/juju/core/instance"
 	network "github.com/juju/juju/core/network"
@@ -595,10 +594,10 @@ func (c *MockMachineSetLinkLayerDevicesCall) DoAndReturn(f func(...state.LinkLay
 }
 
 // Units mocks base method.
-func (m *MockMachine) Units() ([]provisioner.Unit, error) {
+func (m *MockMachine) Units() ([]Unit, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Units")
-	ret0, _ := ret[0].([]provisioner.Unit)
+	ret0, _ := ret[0].([]Unit)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
@@ -616,19 +615,19 @@ type MockMachineUnitsCall struct {
 }
 
 // Return rewrite *gomock.Call.Return
-func (c *MockMachineUnitsCall) Return(arg0 []provisioner.Unit, arg1 error) *MockMachineUnitsCall {
+func (c *MockMachineUnitsCall) Return(arg0 []Unit, arg1 error) *MockMachineUnitsCall {
 	c.Call = c.Call.Return(arg0, arg1)
 	return c
 }
 
 // Do rewrite *gomock.Call.Do
-func (c *MockMachineUnitsCall) Do(f func() ([]provisioner.Unit, error)) *MockMachineUnitsCall {
+func (c *MockMachineUnitsCall) Do(f func() ([]Unit, error)) *MockMachineUnitsCall {
 	c.Call = c.Call.Do(f)
 	return c
 }
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockMachineUnitsCall) DoAndReturn(f func() ([]provisioner.Unit, error)) *MockMachineUnitsCall {
+func (c *MockMachineUnitsCall) DoAndReturn(f func() ([]Unit, error)) *MockMachineUnitsCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
@@ -759,10 +758,10 @@ func (m *MockUnit) EXPECT() *MockUnitMockRecorder {
 }
 
 // Application mocks base method.
-func (m *MockUnit) Application() (provisioner.Application, error) {
+func (m *MockUnit) Application() (Application, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Application")
-	ret0, _ := ret[0].(provisioner.Application)
+	ret0, _ := ret[0].(Application)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
@@ -780,19 +779,19 @@ type MockUnitApplicationCall struct {
 }
 
 // Return rewrite *gomock.Call.Return
-func (c *MockUnitApplicationCall) Return(arg0 provisioner.Application, arg1 error) *MockUnitApplicationCall {
+func (c *MockUnitApplicationCall) Return(arg0 Application, arg1 error) *MockUnitApplicationCall {
 	c.Call = c.Call.Return(arg0, arg1)
 	return c
 }
 
 // Do rewrite *gomock.Call.Do
-func (c *MockUnitApplicationCall) Do(f func() (provisioner.Application, error)) *MockUnitApplicationCall {
+func (c *MockUnitApplicationCall) Do(f func() (Application, error)) *MockUnitApplicationCall {
 	c.Call = c.Call.Do(f)
 	return c
 }
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockUnitApplicationCall) DoAndReturn(f func() (provisioner.Application, error)) *MockUnitApplicationCall {
+func (c *MockUnitApplicationCall) DoAndReturn(f func() (Application, error)) *MockUnitApplicationCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
@@ -859,10 +858,10 @@ func (m *MockApplication) EXPECT() *MockApplicationMockRecorder {
 }
 
 // Charm mocks base method.
-func (m *MockApplication) Charm() (provisioner.Charm, bool, error) {
+func (m *MockApplication) Charm() (Charm, bool, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Charm")
-	ret0, _ := ret[0].(provisioner.Charm)
+	ret0, _ := ret[0].(Charm)
 	ret1, _ := ret[1].(bool)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
@@ -881,19 +880,19 @@ type MockApplicationCharmCall struct {
 }
 
 // Return rewrite *gomock.Call.Return
-func (c *MockApplicationCharmCall) Return(arg0 provisioner.Charm, arg1 bool, arg2 error) *MockApplicationCharmCall {
+func (c *MockApplicationCharmCall) Return(arg0 Charm, arg1 bool, arg2 error) *MockApplicationCharmCall {
 	c.Call = c.Call.Return(arg0, arg1, arg2)
 	return c
 }
 
 // Do rewrite *gomock.Call.Do
-func (c *MockApplicationCharmCall) Do(f func() (provisioner.Charm, bool, error)) *MockApplicationCharmCall {
+func (c *MockApplicationCharmCall) Do(f func() (Charm, bool, error)) *MockApplicationCharmCall {
 	c.Call = c.Call.Do(f)
 	return c
 }
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockApplicationCharmCall) DoAndReturn(f func() (provisioner.Charm, bool, error)) *MockApplicationCharmCall {
+func (c *MockApplicationCharmCall) DoAndReturn(f func() (Charm, bool, error)) *MockApplicationCharmCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
